+++ vrtool/decision_making/measures/standard_measures/vertical_piping_measure_calculators/heavescreen_measure_calculator.py
@@ -20,10 +20,6 @@ class HeavescreenMeasureCalculator(
         * The assumption is that it should go 6m below the cover_layer, so `l_screen = cover_layer_thickness + 6m`.
     """
 
-    @property
-    def pf_piping_reduction_factor(self) -> float:
-        return 1000
-
     def calculate_cost(self, unit_costs: MeasureUnitCosts) -> float:
         _vertical_length = self.dike_section.cover_layer_thickness + 6
         return unit_costs.heavescreen * _vertical_length * self.dike_section.Length