+++ vrtool/decision_making/measures/standard_measures/vertical_piping_measure_calculators/vertical_geotextile_measure_calculator.py
@@ -16,10 +16,6 @@ class VerticalGeotextileMeasureCalculator(
     * It reduces the `pf_piping` with a factor `1000`.
     """
 
-    @property
-    def pf_piping_reduction_factor(self) -> float:
-        return 1000
-
     def calculate_cost(self, unit_costs: MeasureUnitCosts) -> float:
         return unit_costs.vertical_geotextile * self.dike_section.Length