+++ vrtool/orm/models/standard_measure.py
@@ -18,12 +18,10 @@ class StandardMeasure(OrmBaseModel):
     crest_step = FloatField(default=0.5)
     max_crest_increase = FloatField(default=2)
     stability_screen = BooleanField(default=0)
-    prob_of_solution_failure = FloatField(default=1 / 1000)
-    failure_probability_with_solution = FloatField(default=10**-12)
-    stability_screen_s_f_increase = FloatField(default=0.2)
-    transition_level_increase_step = FloatField(default=0.25)
     max_pf_factor_block = FloatField(default=1000)
+    transition_level_increase_step = FloatField(default=0.25)
     n_steps_block = IntegerField(default=4)
+    piping_reduction_factor = FloatField(null=True)
 
     class Meta:
         table_name = _get_table_name(__qualname__)